+++ commons/src/layer.rs
@@ -1,6 +1,6 @@
-pub mod cache_buddy;
 mod configure_env_layer;
 mod default_env_layer;
+pub mod diff_migrate;
 
 #[deprecated(note = "Use the struct layer API in the latest libcnb.rs instead")]
 pub use self::configure_env_layer::ConfigureEnvLayer;
